# apps/be/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/catalog-node-versions/{catalogNodeVersionId}/properties": {
      "get": {
        "operationId": "CatalogNodePropertiesList_findAll",
        "parameters": [
          {
            "name": "catalogNodeVersionId",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Column to sort by (resource-specific whitelist)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "ASC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Text search (maps to searchable columns per endpoint)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isRequired",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCatalogNodeProperties"
                }
              }
            }
          }
        },
        "summary": "List property schemas on a catalog version (paginated)",
        "tags": [
          "Catalog - node properties"
        ]
      }
    },
...
    "/boards": {
      "post": {
        "operationId": "Boards_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBoardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          }
        },
        "summary": "Create board",
        "tags": [
          "Boards"
        ]
      },
      "get": {
        "operationId": "Boards_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Column to sort by (resource-specific whitelist)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "ASC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Text search (maps to searchable columns per endpoint)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedBoards"
                }
              }
            }
          }
        },
        "summary": "List boards (paginated)",
        "tags": [
          "Boards"
        ]
      }
    },
    "/boards/{id}": {
      "get": {
        "operationId": "Boards_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          }
        },
        "summary": "Get board by id",
        "tags": [
          "Boards"
        ]
      },
      "patch": {
        "operationId": "Boards_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          }
        },
        "summary": "Update board (name, snap)",
        "tags": [
          "Boards"
        ]
      },
      "delete": {
        "operationId": "Boards_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        },
        "summary": "Soft-delete board",
        "tags": [
          "Boards"
        ]
      }
    },
    "/boards/{boardId}/nodes": {
      "get": {
        "operationId": "BoardNodesList_findAll",
        "parameters": [
          {
            "name": "boardId",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Column to sort by (resource-specific whitelist)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "ASC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Text search (maps to searchable columns per endpoint)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "catalogNodeVersionId",
            "required": false,
            "in": "query",
            "description": "Filter by pinned catalog version id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "description": "Comma-separated relations: sockets, catalogNodeVersion",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedBoardNodes"
                }
              }
            }
          }
        },
        "summary": "List board nodes for a board (paginated)",
        "tags": [
          "Boards"
        ]
      }
    },
...
    "/board-nodes": {
      "post": {
        "operationId": "BoardNodes_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBoardNodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardNode"
                }
              }
            }
          }
        },
        "summary": "Place a node on a board (pins catalog version)",
        "tags": [
          "Board - nodes"
        ]
      }
    },
    "/board-nodes/{id}": {
      "get": {
        "operationId": "BoardNodes_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardNode"
                }
              }
            }
          }
        },
        "summary": "Get board node by id",
        "tags": [
          "Board - nodes"
        ]
      },
      "patch": {
        "operationId": "BoardNodes_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoardNodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardNode"
                }
              }
            }
          }
        },
        "summary": "Update board node",
        "tags": [
          "Board - nodes"
        ]
      },
      "delete": {
        "operationId": "BoardNodes_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        },
        "summary": "Soft-delete board node",
        "tags": [
          "Board - nodes"
        ]
      }
    },
...
  },
  "info": {
    "title": "Shake API",
    "description": "REST API for the **node catalog** (toolbox definitions) and **board graph** (runtime nodes, connections, props).\n\nEntity IDs are `bigint` in PostgreSQL and serialized as **strings** in JSON.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PaginatedMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20,
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "number",
            "example": 142,
            "minimum": 0
          },
          "totalPages": {
            "type": "number",
            "example": 8,
            "minimum": 0
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "CatalogNodeVersion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "updatedById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "catalogNodeId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "version": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "HTTP Request v1",
            "maxLength": 255
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "deprecatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "catalogNodeId",
          "version",
          "name",
          "isActive",
          "deprecatedAt"
        ]
      },
...
      "CatalogNodeProperty": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "updatedById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "type": {
            "type": "string",
            "example": "string",
            "maxLength": 64
          },
          "defaultValue": {
            "type": "object",
            "nullable": true
          },
          "isRequired": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "catalogNodeVersionId",
          "type",
          "isRequired"
        ]
      },
      "PaginatedCatalogNodeProperties": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogNodeProperty"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "CreateBoardDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "example": "My board"
          },
          "snap": {
            "type": "object",
            "nullable": true,
            "description": "Presentation JSON (viewport, layout); optional",
            "example": {
              "zoom": 1,
              "x": 0,
              "y": 0
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "Board": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "updatedById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string",
            "example": "My board",
            "maxLength": 255
          },
          "snap": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "name"
        ]
      },
      "PaginatedBoards": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Board"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateBoardDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "example": "My board"
          },
          "snap": {
            "type": "object",
            "nullable": true,
            "description": "Presentation JSON (viewport, layout); optional",
            "example": {
              "zoom": 1,
              "x": 0,
              "y": 0
            }
          }
        }
      },
      "BoardNodeSocket": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "updatedById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "boardId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "nodeId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "catalogNodeSocketId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "boardId",
          "nodeId",
          "catalogNodeSocketId"
        ]
      },
      "BoardNode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "updatedById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "boardId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "value": {
            "type": "object",
            "example": "default payload",
            "maxLength": 512,
            "nullable": true
          },
          "sockets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardNodeSocket"
            }
          },
          "catalogNodeVersion": {
            "$ref": "#/components/schemas/CatalogNodeVersion"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "boardId",
          "catalogNodeVersionId"
        ]
      },
      "PaginatedBoardNodes": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardNode"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "CreateBoardNodeDto": {
        "type": "object",
        "properties": {
          "boardId": {
            "type": "string",
            "example": "1",
            "description": "FK to boards.id"
          },
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "Pinned catalog version; FK to catalog_node_versions.id"
          },
          "value": {
            "type": "string",
            "nullable": true,
            "maxLength": 512,
            "example": "instance-meta",
            "description": "Optional runtime value on the placed node"
          }
        },
        "required": [
          "boardId",
          "catalogNodeVersionId"
        ]
      },
      "UpdateBoardNodeDto": {
        "type": "object",
        "properties": {
          "boardId": {
            "type": "string",
            "example": "1",
            "description": "FK to boards.id"
          },
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "Pinned catalog version; FK to catalog_node_versions.id"
          },
          "value": {
            "type": "string",
            "nullable": true,
            "maxLength": 512,
            "example": "instance-meta",
            "description": "Optional runtime value on the placed node"
          }
        }
      },
...
    }
  }
}